.me/extractior/pyproject.toml
[project]
name = "extractior"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "langchain-core>=1.2.14",
    "langchain-openai>=1.1.10",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "pymupdf4llm>=0.3.4",
    "python-dotenv>=1.2.1",
]
